Extend crocodile to call GPU degrid
Extending crocodile to load a GPUDegrid module
and call the GPU-enabled degrid code. To use it, you
must first build the module in the GPUDegrid repository
https://github.com/SKA-ScienceDataProcessor/GPUDegrid
Modify Defines.h so to match the crocodile parameters
for image size, gcf size and the subgrid points between
image points. The build GPUDegrid.so. Make sure PYTHONPATH
includes the location of this file and run crocodile.