Now makes an image correctly
Lots of problems originating in some of the crocodile code having side-effects such as sorting uvw and vis on w. Returning munged uvw is a bad idea anyway because of the necessary scaling by frequency. wcach* do not do anything to w. wslic* still do. And there is a mysterious negating of w floating around. Still the notebook runs and produces a reasonable image.