Phase rotation, third pass
Now all test cases are consistent, so this should really be it. There
was still a sign error with the phasor - we could equivalently calculate
the phasor from the reprojected coordinate system, but that would require
us to determine the rotated UVWs before we can calculate l and m. The
"simulate" test case demonstrates this.

Unfortunately, this requires un-doing the coordinate inconsistencies in
fourier_transforms too, which probably makes yet more test cases fail.