ARL adjustments for crocodile
We can get rid of a lot of strange coordinate transformations now. The
results aren't exactly the same as before (~1% difference), but that
might be coincidence?

Also a bunch of stylistic changes in here - nothing wrong with having
accessors, especially if it means we can save a bunch of variables.