W snapshot notebook
Good news is that it works! At least for small shifts the image stays
good, everything seems to be doing its job.

Bad news is that for large shifts the image gets a lot worse. w-kernels
seem to be going off-center more and more until they are effectively
useless. Hm.