Hermitian grid optimisation
Yet another thing I wanted to have a closer look at. The way it looks
this might actually be what we got wrong in the RC prototype: For making
a grid "hermitian" we need to mirror around the zero frequency, which for
FFT is at N//2. For even size this leads to the usual asymmetry.