Switch to "positive" oversampling coordinates
Sample oversampled values at positive sub-grid values. This means that
using a certain sub-grid index now is equivalent to moving into a
*negative* direction.

Why this change? It is the more logical order for oversampling values,
and especially makes more sense for kernels with even support: As the
zero axis is half a cell to the right, this means we (over)sample from
-N/2+o to N/2-1+o (with o in [0;1[ the sub-cell offset). This is exactly
symetrical, in contrast to what we had before (-N/2-o to N/2-1-o). The
situation doesn't change for uneven supports.

Normally imaging frameworks just flip the axes at this point to make the
coordinates easier, but this is actually not acceptable to us, as we are
using asymmetrical kernels quite a bit. So we are doing it the hard
way instead.