Notebooks for A-imaging, predict
This (finally!) adds a proper explanation for what is going on with the
A-terms. This is still simplified as it ignores the actual primary beam
as well as all time and frequency dependency. But this should get the
core of the problem across rather well.

Also added a notebook for checking predict accuracy, and moved the
wsnapshot-new notebook to wtowers.