Introduce facet/subgrid spacings
Proper way to reason about valid facet/subgrid centres. We should be able
to transition away from using xA now. After all, this ought to just refer
to the amount of subgrid data that is usable with respect to all the
convolutions applied, which the recombination code itself doesn't need
to care about.