Finishing with configuration work
We now split subgrids that are too big - likely no way around that feature,
even though BLDA should make it less impactful at some point. In reality
it shouldn't matter, we can afford to transfer a couple of subgrids to
more nodes than strictly needed. In fact, this is more often than not
exactly the central subgrid, which will be worked on by so many workers
that it might be worth broadcasting! Future micro-optimisation...

Now actually does proper chunked visibility I/O, just not hooked up to
anything yet. At this point test_config can already act as a bit of an
HDF5 benchmark. Nothing really interesting yet, and the numbers are not
reliable as we currently re-write and re-read a bunch of chunks
unnecessarily.