Skip receive for duplicated subgrid transfers
We were already skipping work, but clearly we should not attempt to
receive something that isn't going to arrive. Work distribution was
changed so duplicated transfers only happen rarely, which means that this
manifested as rare hangs for particular configurations - but we should
really be robust against this sort of thing.