Fix nasty race condition for subgrids sent to multiple streamers
Another one of those obvious mistakes that take forever to find because
they matter so rarely. In this case, this only applies to sugrids that
get sent to multiple streamers (generally only subgrid 0/0), and then
requires the send slot of the first copy to get overwritten while the
second has not started yet. Likely was masked before the scheduling
changes because subgrid 0/0 was always sent first.