Input thread starts doing work if there's enough tasks available
No need to block an entire thread for waiting for more input from MPI
if there's already more than enough work to do. This is clearly
slightly risky - if we do this too eagerly, the receive thread taking
on work might cause us to run out of tasks to run. Need to consider
this in relation to the baselines-per-task parameter.